% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbbreviationPipe.R
\docType{class}
\name{AbbreviationPipe}
\alias{AbbreviationPipe}
\title{Class to find and/or replace the abbreviations on the data of an instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbbreviationPipe$new(propertyName = "abbreviation",
                     propertyLanguageName = "language",
                     pathResourcesAbbreviations = "resources/abbreviations-json",
                     alwaysBeforeDeps = list("GuessLanguagePipe"),
                     notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{propertyLanguageName}{(character) Name of the language property.}

\item{pathResourcesAbbreviations}{(character) Path where are stored the}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the abbreviations that are in it. Optionally, you can decide whether to
replace the data abbreviations or not.
}
\details{
This class needs files in json format that will contain the
abbreviations to be located and the string that will replace them. For this,
it is necessary that the instance contains a property that indicates the
language of the data to be able to correctly choose the list of abbreviations
that apply to the data. The format of the file names of the resources has to
be: abbrev.xxx.json (Being xxx the value of the language property of the
instance).

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/replace the abbreviations.
The abbreviations found in the pipe are added to the list of properties of
the Instance. If the replaceAbbreviations parameter is TRUE, the instance
data will be modified by replacing the abbreviations found.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, replaceAbbreviations = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{replaceAbbreviations}}{
(logical) Indicate if the abbreviations are replaced or not.
}
}
}
}
}

\item{\bold{findAbbreviation}}{
Function that checks if the abbreviation is in the data.
\itemize{
\item{\emph{Usage}}{

\code{findAbbreviation(data, abbreviation)}
}
\item{\emph{Value}}{

TRUE or FALSE depending on whether the abbreviation is in the data.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which the abbreviation is searched.
}
\item{\strong{abbreviation}}{
(character) Indicates the abbreviation to find.
}
}
}
}
}

\item{\bold{replaceAbbreviation}}{
Function that replace the abbreviation in the data for the extendedAbbreviation.
\itemize{
\item{\emph{Usage}}{

\code{replaceAbbreviation(abbreviation, extendedAbbreviation, data)}
}
\item{\emph{Value}}{

The data with the abbreviatons replaced.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{abbreviation}}{
(character) Indicates the abbreviation to replace.
}
\item{\strong{extendedAbbreviation}}{
(character) Indicates the string to replace for the abbreviations found.
}
\item{\strong{data}}{
(character) Text in which abbreviations will be replaced.
}
}
}
}
}

\item{\bold{getPropertyLanguageName}}{
Getter of name of property language.
\itemize{
\item{\emph{Usage}}{

\code{getPropertyLanguageName()}
}
\item{\emph{Value}}{

Value of name of property language.
}
}
}

\item{\bold{getPathResourcesAbbreviations}}{
Getter of path of abbreviations resources.
\itemize{
\item{\emph{Usage}}{

\code{getPathResourcesAbbreviations()}
}
\item{\emph{Value}}{

Value of path of abbreviations resources.
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{propertyLanguageName}}{
 (character) The name of property about language.
}
\item{\bold{pathResourcesAbbreviations}}{
 (character) The path where are the resources.
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}},
\code{\link{ResourceHandler}}
}
\keyword{datasets}
