% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BabelfyEntry.R
\docType{class}
\name{BabelfyEntry}
\alias{BabelfyEntry}
\title{Class to represent a babelfy Semantic annotation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BabelfyEntry$new(startIdx, endIdx, score, synsetId, text)
}
\arguments{
\item{startIdx}{(numeric) The start index of an entry.}

\item{endIdx}{(numeric) The last index of the entrys.}

\item{score}{(numeric) The score.}

\item{synsetId}{(character) The synset ID.}

\item{text}{(character) The text of an entry.}
}
\description{
This class is to represent a babelfy Semantic annotation with all relevant
attributes to made intensive searches and discard the irrelevant information
achieved by Babelfy.
}
\section{Methods}{

\itemize{
\item{\bold{getStartIdx}}{
Getter of startIdx variable.
\itemize{
\item{\emph{Usage}}{

\code{getStartIdx()}
}
\item{\emph{Value}}{

Value of startIdx variable.
}
}
}

\item{\bold{setStartIdx}}{
Setter of startIdx variable.
\itemize{
\item{\emph{Usage}}{

\code{setStartIdx(startIdx)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{startIdx}}{
(numeric) The new value of startIdx variable.
}
}
}
}
}

\item{\bold{getEndIdx}}{
Getter of endIdx variable.
\itemize{
\item{\emph{Usage}}{

\code{getEndIdx()}
}
\item{\emph{Value}}{

Value of endIdx variable.
}
}
}

\item{\bold{setEndIdx}}{
Setter of endIdx variable.
\itemize{
\item{\emph{Usage}}{

\code{setEndIdx(endIdx)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{endIdx}}{
(numeric) The new value of endIdx variable.
}
}
}
}
}

\item{\bold{getScore}}{
Getter of score variable.
\itemize{
\item{\emph{Usage}}{

\code{getScore()}
}
\item{\emph{Value}}{

Value of score variable.
}
}
}

\item{\bold{setScore}}{
Setter of score variable.
\itemize{
\item{\emph{Usage}}{

\code{setScore(score)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{score}}{
(numeric) The new value of score variable.
}
}
}
}
}

\item{\bold{getSynsetId}}{
Getter of synsetId variable.
\itemize{
\item{\emph{Usage}}{

\code{getSynsetId()}
}
\item{\emph{Value}}{

Value of synsetId variable.
}
}
}

\item{\bold{setSynsetId}}{
Setter of synsetId variable.
\itemize{
\item{\emph{Usage}}{

\code{setSynsetId(synsetId)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsetId}}{
(character) The new value of synsetId variable.
}
}
}
}
}

\item{\bold{getText}}{
Getter of text variable.
\itemize{
\item{\emph{Usage}}{

\code{getText()}
}
\item{\emph{Value}}{

Value of text variable.
}
}
}

\item{\bold{setText}}{
Setter of text variable.
\itemize{
\item{\emph{Usage}}{

\code{setText(text)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{text}}{
(character) The new value of text variable.
}
}
}
}
}

}
}

\keyword{datasets}
