% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connections.R
\docType{class}
\name{Connections}
\alias{Connections}
\title{Class to manage the connections with twitter and youtube}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Connections$new(pathKeys)
}
\arguments{
\item{pathKeys}{(character) Path of the .ini file that contains the keys.}
}
\description{
The tasks of the functions that the Connections class has are to
establish the connections and control the number of requests that have been made.
}
\details{
The way to indicate the keys of youtube and twitter has to be
through an .ini file that contains the following structure:

[twitter]

ConsumerKey=YourConsumerKey

ConsumerSecret=YourConsumerSecret

AccessToken=YourAccessToken

AccessTokenSecret=YourAccessTokenSecret

[youtube]

app_id=YourAppId

app_password=YourAppPassword
}
\section{Methods}{

\itemize{
\item{\bold{getTwitterToken}}{
Getter of twitterToken.
\itemize{
\item{\emph{Usage}}{

\code{getTwitterToken()}
}
\item{\emph{Value}}{

Value of twitterToken.
}
}
}

\item{\bold{startConnectionWithTwitter}}{
Function that establishes the connection to twitter. If the connection has
not been established, the keys necessary to make the connection are indicated
in the setup_twitter_oauth function. Then it is indicated that the connection
has been established.
\itemize{
\item{\emph{Usage}}{

\code{startConnectionWithTwitter()}
}
}
}

\item{\bold{checkRequestToTwitter}}{
Function that controls the connection with twitter. If the limit of twitter
requests has been exceeded, 15 minutes are expected.
\itemize{
\item{\emph{Usage}}{

\code{checkRequestToTwitter()}
}
}
}

\item{\bold{startConnectionWithYoutube}}{
Function that establishes the connection to youtube. If the connection has
not been established, the keys necessary to make the connection are indicated
in the yt_oauth function. Then it is indicated that the connection has been
established.
\itemize{
\item{\emph{Usage}}{

\code{startConnectionWithYoutube()}
}
}
}

\item{\bold{addNumRequestToYoutube}}{
Function that increases in one the number of request to youtube.
\itemize{
\item{\emph{Usage}}{

\code{addNumRequestToYoutube()}
}
}
}

\item{\bold{checkRequestToYoutube}}{
Function that controls the connection with youtube. If the limit of youtube
requests has been exceeded, 15 minutes are expected.
\itemize{
\item{\emph{Usage}}{

\code{checkRequestToYoutube()}
}
}
}

\item{\bold{getNumRequestMaxToYoutube}}{
Getter of num max of from Youtube.
\itemize{
\item{\emph{Usage}}{

\code{getNumRequestMaxToYoutube()}
}
\item{\emph{Value}}{

Value of num max of from Youtube.
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{keys}}{
 (list) The keys of twitter and youtube.
}
\item{\bold{numRequestToYoutube}}{
 (numeric) Indicates the number of requests made.
}
\item{\bold{numRequestMaxToYoutube}}{
 (numeric) Indicates the maximum number of requests with Youtube.
}
\item{\bold{connectionWithYoutube}}{
 (logical) Indicates if the connection has been established with Youtube.
}
\item{\bold{connectionWithTwitter}}{
 (logical) Indicates if the connection has been established with Twitter.
}
\item{\bold{twitterToken}}{
 (Token) Token to establish the connection to twitter.
}
}
}

\seealso{
\code{\link{ExtractorTwtid}}, \code{\link{ExtractorYtbid}}
}
\keyword{datasets}
