% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorEml.R
\docType{class}
\name{ExtractorEml}
\alias{ExtractorEml}
\title{Class to handle eml files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorSms$new(path)
}
\arguments{
\item{path}{(character) Path of the eml-type file.}
}
\description{
It is a class that inherits from the \code{Instance} class and
implements the functions of extracting the text and the date of an eml-type
file.
}
\details{
The way to indicate which part to choose in the email is through an
.ini file which contains the following structure (being xxxx, text / plain or
text / html):

[eml]

PartSelectedOnMPAlternative = xxxx

To be able to use this class it is necessary to have python installed.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainDate}}{
Function that obtain the date of the eml file. Call the function read_emails
and obtain the date of the file indicated in the path and then transforms it
into the generic date format, that is "\%a \%b \%d \%H:\%M:\%S \%Z \%Y"
(Example: "Thu May 02 06:52:36 UTC 2013").
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtains the source of the eml file. Calls the function read_emails
and obtains the source of the file indicated in the path. In addition, it
initializes the data with the initial source.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}

\item{\bold{getPartSelectedOnMPAlternative}}{
Getter of PartSelectedOnMPAlternative variable.
\itemize{
\item{\emph{Usage}}{

\code{getPartSelectedOnMPAlternative()}
}
\item{\emph{Value}}{

Value of PartSelectedOnMPAlternative variable.
}
}
}

\item{\bold{setPartSelectedOnMPAlternative}}{
Setter of PartSelectedOnMPAlternative variable.
\itemize{
\item{\emph{Usage}}{

\code{setPartSelectedOnMPAlternative(PartSelectedOnMPAlternative)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{PartSelectedOnMPAlternative}}{
(character) The new value of PartSelectedOnMPAlternative variable.
}
}
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{PartSelectedOnMPAlternative}}{
 (character) Configuration to read the eml files. Indicates whether the
 text/plain part or the text/html part is read.
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
