% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorSms.R
\docType{class}
\name{ExtractorSms}
\alias{ExtractorSms}
\title{Class to handle sms files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorSms$new(path)
}
\arguments{
\item{path}{(character) Path of the sms-type file.}
}
\description{
It is a class that inherits from the Instance class and implements
the functions of extracting the text and the date of an sms-type file.
}
\details{
Due to the fact that the creation date of the message can not be
extracted from the text of an SMS, the date will be initialized to empty.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainDate}}{
Function that obtain the date of the sms file.
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtain the source of the sms file. Reads the file indicated in
the path. In addition it initializes the data with the initial source.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
