% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorTwtid.R
\docType{class}
\name{ExtractorTwtid}
\alias{ExtractorTwtid}
\title{Class to handle twtid files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorTwtid$new(path)
}
\arguments{
\item{path}{(character) Path of the twtid-type file.}
}
\description{
It is a class that inherits from the Instance class and implements
the functions of extracting the text and the date of an twtid-type file.
}
\details{
The connection to twitter is handled through the Connections class,
which needs a configuration file with the necessary keys to make requests to
the twitter API.

This class stores in the folder testFiles/cache/hsspam14/tweets the
tweets processed so far, thus allowing you to save twitter queries. The text
fields, the date and the language of the tweet are stored.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainId}}{
Function that obtain the id of the twtid. Read the id of the file indicated
in the variable path.
\itemize{
\item{\emph{Usage}}{

\code{obtainId()}
}
}
}

\item{\bold{getId}}{
Getter of tweet id.
\itemize{
\item{\emph{Usage}}{

\code{getId()}
}
\item{\emph{Value}}{

Value of tweet id.
}
}
}

\item{\bold{obtainDate}}{
Function that obtain the date of the twtid id. Check if the tweet has
previously been cached. In this case, the file is read in json format and the
date is stored. Otherwise, the request is made on twitter. The date is then
formatted to "\%a \%b \%d \%H:\%M:\%S \%Z \%Y"
(Example: "Thu May 02 06:52:36 UTC 2013").
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtain the source of the twtid id. Check if the tweet has
previously been cached. In this case, the file is read in json format and the
source is stored. Otherwise, the request is made on twitter.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{id}}{
 (character) Id of tweet.
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
