% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorTytb.R
\docType{class}
\name{ExtractorTytb}
\alias{ExtractorTytb}
\title{Class to handle tytb files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorTytb$new(path)
}
\arguments{
\item{path}{(character) Path of the tytb-type file.}
}
\description{
It is a class that inherits from the \code{Instance} class and
implements the functions of extracting the text and the date of an Tytb-type
file.
}
\details{
Due to the fact that the creation date of the message can not be
extracted from the text of an comment from youtube, the date will be
initialized to empty.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainDate}}{
Function that obtains the date of the tytb file.
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtains the source of the tytb file. Reads the file indicated in
the path. In addition, it initializes the data with the initial source.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
