% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorWarc.R
\docType{class}
\name{ExtractorWarc}
\alias{ExtractorWarc}
\title{Class to handle warc files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorWarc$new(path)
}
\arguments{
\item{path}{(character) Path of the warc-type file.}
}
\description{
It is a class that inherits from the \code{Instance} class and
implements the functions of extracting the text and the date of an warc-type
file.
}
\details{
The read_warc function of the jwart package was overwritten because
it returned the hours wrong.

The jwart package makes calls to Java so it is necessary to have rJava
installed.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainDate}}{
Function that obtains the date of the warc file. Finds the warcinfo type records
in which the date appearss and standardizes it with the format:
"\%a \%b \%d \%H:\%M:\%S \%Z \%Y" (Example: "Thu May 02 06:52:36 UTC 2013").
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtains the source of the warc file. The list of records that
contain information are obtained, which they are resource and response. Then
they are traversed and the charset of that record is obtained. If that charset
matches the one obtained from guess_encoding, payload_content is used to get
the contents of the record. If it does not match, the content is obtained,
converting the content in bytes to string. This is done because there are
coding problems in cases that the charset that is detected is different from
the one that is really. In addition it initializes the data with the initial
source.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}

}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
