% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractorYtbid.R
\docType{class}
\name{ExtractorYtbid}
\alias{ExtractorYtbid}
\title{Class to handle ytbid files}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExtractorTwtid$new(path)
}
\arguments{
\item{path}{(character) Path of the ytbid-type file.}
}
\description{
It is a class that inherits from the \code{Instance} class and
implements the functions of extracting the text and the date of an ytbid-type
file.
}
\details{
The connection to twitter is handled through the \code{Connections}
class, which needs a configuration file with the necessary keys to make
requests to the youtube API.

This class stores in the folder testFiles/cache/youtube/comments the
comments processed so far, thus allowing you to save youtuve queries.
The text fields and the date of the comment are stored.
}
\section{Inherit}{

This class inherits from \code{\link{Instance}} and implements the
\code{obtainSource} and \code{obtainDate} abstracts functions.
}

\section{Methods}{

\itemize{
\item{\bold{obtainId}}{
Function that obtains the id of the ytbid. Read the id of the file indicated
in the variable path.
\itemize{
\item{\emph{Usage}}{

\code{obtainId()}
}
}
}

\item{\bold{getId}}{
Getter of comment id.
\itemize{
\item{\emph{Usage}}{

\code{getId()}
}
\item{\emph{Value}}{

Value of comment id.
}
}
}

\item{\bold{obtainDate}}{
Function that obtains the date of the ytbid id. Check if the comment has
previously been cached. In this case, the file is read in json format and the
date is stored. Otherwise, the request is made on youtube The date is then
formatted to the established standard.
\itemize{
\item{\emph{Usage}}{

\code{obtainDate()}
}
}
}

\item{\bold{obtainSource}}{
Function that obtains the source of the twtid id. Check if the tweet has
previously been cached. In this case, the file is read in json format and the
source is stored. Otherwise, the request is made on twitter.
\itemize{
\item{\emph{Usage}}{

\code{obtainSource()}
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{id}}{
 (character) id of comment.
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
