% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File2Pipe.R
\docType{class}
\name{File2Pipe}
\alias{File2Pipe}
\title{Class to obtain the source of an instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
File2Pipe$new(propertyName = "source",
              alwaysBeforeDeps = list("TargetAssigningPipe"),
              notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
The method of obtaining source is called which implement the
subclasses of the superclass \code{Instance}.
}
\details{
In the case that the source obtained is empty or is not utf-8,
the instance is invalidated.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain the source.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
