% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindEmojiPipe.R
\docType{class}
\name{FindEmojiPipe}
\alias{FindEmojiPipe}
\title{Class to find and/or replace the emoji on the data of an instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindEmojiPipe$new(propertyName = "Emojis",
                  alwaysBeforeDeps = list(),
                  notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the emojis that are in it. Optionally, you can decide whether to
replace the data emojis or not.
}
\details{
The emoji list that is used is the one provided by the variable
\code{emojis} of the rtweet package.

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/replace the abbreviations.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, replaceEmoji = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{replaceEmoji}}{
(logical) Indicate if the emojis are replaced.
}
}
}
}
}

\item{\bold{findEmoji}}{
Function that checks if the emoji is in the data.
\itemize{
\item{\emph{Usage}}{

\code{findEmoji(data, emoji)}
}
\item{\emph{Value}}{

TRUE or FALSE depending on whether the emoji is on the data.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which the emoji is searched.
}
\item{\strong{emoji}}{
(character) Indicates the emoji to find.
}
}
}
}
}

\item{\bold{replaceEmoji}}{
Function that replaces the emoji in the data for the extendedEmoji.
\itemize{
\item{\emph{Usage}}{

\code{replaceEmoji(emoji, extendedEmoji, data)}
}
\item{\emph{Value}}{

The data with emoji replaced.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{emoji}}{
(character) Indicates the emoji to remove.
}
\item{\strong{extendedEmoji}}{
(character) Indicates the string to replace for the emoji found.
}
\item{\strong{data}}{
(character) Text in which emojis will be replaced.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
