% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindEmoticonPipe.R
\docType{class}
\name{FindEmoticonPipe}
\alias{FindEmoticonPipe}
\title{Class to find and/or remove the emoticon on the data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindEmoticonPipe$new(propertyName = "Emoticon",
                     alwaysBeforeDeps = list(),
                     notAfterDeps = list("FindHashtagPipe"))
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the emoticons that are in it. Optionally, you can decide whether to
remove the data emoticons or not.
}
\details{
The regular expression indicated in the \code{emoticonPattern}
variable is used to identify emoticons.

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/remove the emoticons.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, removeEmoticon = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{removeEmoticon}}{
(logical) Indicate if the emoticons are replaced.
}
}
}
}
}

\item{\bold{findEmoticon}}{
Function that find the emoticons in the data.
\itemize{
\item{\emph{Usage}}{

\code{findEmoticon(data)}
}
\item{\emph{Value}}{

List with emoticons found.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to preproccess.
}
}
}
}
}

\item{\bold{replaceEmoticon}}{
Function that removes the emoticons in the data.
\itemize{
\item{\emph{Usage}}{

\code{replaceEmoticon(data)}
}
\item{\emph{Value}}{

The data with emoticons removed.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which emoticons will be removed.
}
}
}
}
}
}
}

\section{Public fields}{

\itemize{
\item{\bold{emoticonPattern}}{
 (character) Regular expression to detect emoticons.
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
