% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindHashtagPipe.R
\docType{class}
\name{FindHashtagPipe}
\alias{FindHashtagPipe}
\title{Class to find and/or remove the hashtags on the data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindHashtagPipe$new(propertyName = "hashtag",
                     alwaysBeforeDeps = list(),
                     notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the hashtags that are in it. Optionally, you can decide whether to
remove the data hashtags or not.
}
\details{
The regular expression indicated in the \code{hashtagPattern}
variable is used to identify hashtags.

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/remove the hashtags.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, removeHashtag = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{removeHashtag}}{
(logical) Indicates if the hashstags are replaced.
}
}
}
}
}

\item{\bold{findHashtag}}{
Function that find the hashtags in the data.
\itemize{
\item{\emph{Usage}}{

\code{findHashtag(data)}
}
\item{\emph{Value}}{

List with hashtags found.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which the hashtags are searched.
}
}
}
}
}

\item{\bold{removeHashtag}}{
Function that removes the hashtags in the data.
\itemize{
\item{\emph{Usage}}{

\code{removeHashtag(data)}
}
\item{\emph{Value}}{

The data with hashtags removed.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which hashtags will be removed.
}
}
}
}
}
}
}

\section{Public fields}{

\itemize{
\item{\bold{hashtagPattern}}{
 (character) Regular expression to detect hashtags.
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
