% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUserNamePipe.R
\docType{class}
\name{FindUserNamePipe}
\alias{FindUserNamePipe}
\title{Class to find and/or remove the users on the data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindUserNamePipe$new(propertyName = "userName",
                     alwaysBeforeDeps = list(),
                     notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the user names that are in it. Optionally, you can decide whether to
remove the data user names or not.
}
\details{
The regular expressions indicated in the \code{userPattern}
variable are used to identify user names.

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/remove the name users.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, removeUser = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{removeUser}}{
(logical) Indicate if the users are removed.
}
}
}
}
}

\item{\bold{findUserName}}{
Function that find the users in the data.
\itemize{
\item{\emph{Usage}}{

\code{findHashtag(data)}
}
\item{\emph{Value}}{

List with users found.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to preproccess.
}
}
}
}
}

\item{\bold{removeUserName}}{
Function that removes the users in the data.
\itemize{
\item{\emph{Usage}}{

\code{removeUserName(data)}
}
\item{\emph{Value}}{

The data with users removed.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which name users will be removed.
}
}
}
}
}
}
}

\section{Public fields}{

\itemize{
\item{\bold{userPattern}}{
 (character) Regular expression to detect users.
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
