% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeOperator.R
\docType{methods}
\name{freduce}
\alias{freduce}
\title{Apply to the list of functions sequentially and control if the
       instance is invalid}
\usage{
freduce(value, function_list)
}
\arguments{
\item{value}{Initial value.}

\item{function_list}{A list of functions.}
}
\value{
The result after applying each function in turn.
}
\description{
This function applies the first function to value, then the next
             function to the result of the previous function call, etc.
             Execute the functions and allow interruption if the instance is
             invalid.
}
\details{
This is the freduce method of the modified magrittr library in order
         to stop the flow when the instance is invalid.
}
\seealso{
\code{\link{Instance}}
}
