% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeOperator.R
\docType{methods}
\name{\%>|\%}
\alias{\%>|\%}
\title{bdp4R forward-pipe operator}
\usage{
lhs \%>|\% rhs
}
\arguments{
\item{lhs}{An instance object.}

\item{rhs}{A function call using the bdp4R semantics.}
}
\value{
The instance modified by the methods it has traversed.
}
\description{
Function that defines the operator that allows to take the flow
of the preprocessing of the instances.
}
\details{
This is the %>% operator of the modified magrittr library in order
to stop the flow when the instance is invalid.
}
\seealso{
\code{\link{Instance}}
}
