% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessDatePipe.R
\docType{class}
\name{GuessDatePipe}
\alias{GuessDatePipe}
\title{Class to obtain the date}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GuessDatePipe$new(propertyName = "date",
                  alwaysBeforeDeps = list("TargetAssigningPipe"),
                  notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
The method of obtaining date is called which implement the
subclasses of the superclass Instance.
}
\section{Inherit}{

This class inherit from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain the date.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
