% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuessLanguagePipe.R
\docType{class}
\name{GuessLanguagePipe}
\alias{GuessLanguagePipe}
\title{Class to guess the language}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GuessLanguagePipe$new(propertyName = "language",
                      alwaysBeforeDeps = list("StoreFileExtensionPipe",
                                              "TargetAssigningPipe",
                                              "StripHTMLPipe"),
                      notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Class to guess the language by using language detector of library
cld2. Optionally, it is possible to choose the language provided by twitter.
}
\details{
The pipe will invalidate the instance if the language of the data
can not be detect.

To obtain the language of the tweets, it will be verified that there is a
json file with the information stored in memory.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain the language of the data.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, languageTwitter = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{languageTwitter}}{
(logical) Indicates whether for the instances of type twtid the language that
returns the api is obtained or the detector is applied.
}
}
}
}
}

\item{\bold{getLanguage}}{
Function that guess the language of data.
\itemize{
\item{\emph{Usage}}{

\code{getLanguage(data)}
}
\item{\emph{Value}}{

The language guesser. Format: see ISO 639-3:2007.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to guess the language.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
