% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstanceFactory.R
\docType{class}
\name{InstanceFactory}
\alias{InstanceFactory}
\title{Class to build Instance types}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
InstanceFactory$new()
}
\description{
This class uses the standard factory method because it is
responsible for choosing the type of instances that will be created from the
files indicated in the path.
}
\section{Methods}{

\itemize{
\item{\bold{createInstance}}{
Function that builds instances from the path it receives. Depending on the
extension of the file indicated in the path, an instance type or other is
created.
\itemize{
\item{\emph{Usage}}{

\code{createInstance(path)}
}
\item{\emph{Value}}{

The corresponding object according to the file extension.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{path}}{
(character) Path of the file to create an instance.
}
}
}
}
}
}
}

\seealso{
\code{\link{Instance}}, \code{\link{ExtractorEml}},
\code{\link{ExtractorSms}}, \code{\link{ExtractorTtwt}},
\code{\link{ExtractorTwtid}}, \code{\link{ExtractorTytb}},
\code{\link{ExtractorWarc}}, \code{\link{ExtractorYtbid}}
}
\keyword{datasets}
