% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureLengthPipe.R
\docType{class}
\name{MeasureLengthPipe}
\alias{MeasureLengthPipe}
\title{Class to obtain the length of the data of an instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MeasureLengthPipe$new(propertyName = "length",
                      alwaysBeforeDeps = list(),
                      notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to know the length of the data of an
instance.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain the length of data.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, propertyName = super$getPropertyName(),
nchar_conf = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{propertyName}}{
(character) The name of the property that will be obtained in the pipe.
}
\item{\strong{nchar_conf}}{
(logical)
}
}
}
}
}

\item{\bold{getLength}}{
Function that obtain the length of the data.
\itemize{
\item{\emph{Usage}}{

\code{getLength(data, nchar_conf = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to preproccess.
}
\item{\strong{nchar_conf}}{
(logical)
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
