% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeGeneric.R
\docType{class}
\name{PipeGeneric}
\alias{PipeGeneric}
\title{Super class that handles the general functionalities of the management
of the pipes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PipeGeneric$new(propertyName,
                alwaysBeforeDeps,
                notAfterDeps)
}
\arguments{
\item{propertyName}{(character) Name of the property.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Super class that handles the general functionalities of the
pipes.
}
\details{
Building...
}
\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Abtract method to preprocess the intance.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preprocess.
}
}
}
}
}

\item{\bold{getPropertyName}}{
Getter of name of property.
\itemize{
\item{\emph{Usage}}{

\code{getPropertyName()}
}
\item{\emph{Value}}{

Value of name of property.
}
}
}

\item{\bold{getAlwaysBeforeDeps}}{
Getter of the dependences always before.
\itemize{
\item{\emph{Usage}}{

\code{getAlwaysBeforeDeps()}
}
\item{\emph{Value}}{

Value of dependences always before.
}
}
}

\item{\bold{getNotAfterDeps}}{
Getter of the dependences not after.
\itemize{
\item{\emph{Usage}}{

\code{getNotAfterDeps()}
}
\item{\emph{Value}}{

Value of dependences not after.
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{propertyName}}{
 (character) The name of property.
}
\item{\bold{alwaysBeforeDeps}}{
 (list) Dependencies of the type alwaysBefore. These dependences indicate
 what pipes must be executed before the current one.
}
\item{\bold{notAfterDeps}}{
 (list) Dependencies of the type notAfter. These dependences indicate what
 pipes must not be executed after the current one.
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
