% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_warc.R
\docType{methods}
\name{read_warc}
\alias{read_warc}
\title{Read a WARC file (compressed or uncompressed)}
\usage{
read_warc(path, warc_types = NULL, include_payload = FALSE)
}
\arguments{
\item{path}{Path to WARF file.}

\item{warc_types}{If not NULL and one or more of warcinfo, request, response,
resource, metadata, revisit, conversion then returned WARC
records will be filtered to only include the specified
record types.}

\item{include_payload}{If TRUE then the payload for each WARC record will be
included.}
}
\value{
listInstances list of instances that have been preprocessed
}
\description{
Method that allows to process files in a comfortable way.
}
\details{
Function of the jwart package that has been modified to read correctly
         the date of the warc records. This override has been made since the
         date obtained was without the time because the transformation of the
         format of the warc dates to the standard format was incompatible.
}
\examples{
read_warc(system.file("extdata/bbc.warc", package="jwatr"))
read_warc(system.file("extdata/sample.warc.gz", package="jwatr"),
          warc_types = "response", include_payload = FALSE)


}
\seealso{
\code{\link{ExtractorWarc}}
}
