% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerialPipes.R
\docType{class}
\name{SerialPipes}
\alias{SerialPipes}
\title{Class that implements the flow of pipes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SerialPipes$new()
}
\description{
This class inherits from the TypePipe class, which has the
pipeAll method that has a default implementation.
}
\details{
This class uses the default implementation provided by its parent
class and manages unexpected errors that may appear in the execution of the
pipes.
}
\section{Inherit}{

This class inherits from \code{\link{TypePipe}} and implements the
\code{pipeAll} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipeAll}}{
Function that implements the pipeAll function of the super class. In this
case, the pipeAll function is called so that the default pipe stream is
executed
\itemize{
\item{\emph{Usage}}{

\code{pipeAll(instance)}
}
\item{\emph{Value}}{

The preprocessed instance.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) The instance that is going to be processed.
}
}
}
}
}

}
}

\seealso{
\code{\link{TypePipe}}, \code{\link{Instance}}
}
\keyword{datasets}
