% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StopWordPipe.R
\docType{class}
\name{StopWordPipe}
\alias{StopWordPipe}
\title{Class to find and/or remove the stop words on the data of an instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StopWordPipe$new(propertyName = "stopWord",
                 propertyLanguageName = "language",
                 alwaysBeforeDeps = list("GuessLanguagePipe"),
                 notAfterDeps = list("AbbreviationPipe"))
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{propertyLanguageName}{(character) Name of the language property.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
This class allows you to preprocess the data of an instance to
find the stop words that are in it. Optionally, you can decide whether to
remove the data stop words or not.
}
\details{
This class needs files in json format that will contain the
stop words to be located. For this it is necessary that the instance contains
a property that indicates the language of the data to be able to correctly
choose the list of stop words that apply to the data. The format of the file
names of the resources has to be: xxx.json (Being xxx the value of the
language property of the instance).

To indicate the path where the associated resources are located, the
configuration file is used. It is necessary to indicate in the section called
resourcesPath, the path of resourcesStopWordsPath.

The pipe will invalidate the instance in the moment that the resulting data is
empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain/remove the stop words.
The stop words found in the pipe are added to the list of properties of
the Instance. If the removeStopWords parameter is TRUE, the instance data
will be removed.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, removeStopWords = TRUE)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{removeStopWords}}{
(logical) Indicates if the stop words are removed or not.
}
}
}
}
}

\item{\bold{findStopWord}}{
Function that checks if the stop word is in the data.
\itemize{
\item{\emph{Usage}}{

\code{findStopWord(data, stopWord)}
}
\item{\emph{Value}}{

TRUE or FALSE depending on whether the stop word is on the data.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text in which the stop word is searched.
}
\item{\strong{stopWord}}{
(character) Indicates the stop word to find.
}
}
}
}
}

\item{\bold{removeStopWord}}{
Function that removes the stop word in the data.
\itemize{
\item{\emph{Usage}}{

\code{removeStopWord(stopWord, data)}
}
\item{\emph{Value}}{

The data with stop word removed.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{stopWord}}{
(character) Indicates the stop word to remove.
}
\item{\strong{data}}{
(character) Text in which stop words will be removed.
}
}
}
}
}

\item{\bold{getPropertyLanguageName}}{
Getter of name of property language.
\itemize{
\item{\emph{Usage}}{

\code{getPropertyLanguageName()}
}
\item{\emph{Value}}{

Value of name of property language.
}
}
}

\item{\bold{getPathResourcesStopWords}}{
Getter of path of stop words resources.
\itemize{
\item{\emph{Usage}}{

\code{getPathResourcesStopWords()}
}
\item{\emph{Value}}{

Value of path of stop words resources.
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{propertyLanguageName}}{
 (character) The name of property about language.
}
\item{\bold{pathResourcesStopWords}}{
 (character) The path where are the resources.
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}},
\code{\link{ResourceHandler}}
}
\keyword{datasets}
