% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoreFileExtensionPipe.R
\docType{class}
\name{StoreFileExtensionPipe}
\alias{StoreFileExtensionPipe}
\title{Class to get the extension of a file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StoreFileExtensionPipe$new(propertyName = "extension",
                           alwaysBeforeDeps = list(""),
                           notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Class to get the extension of a file.
}
\details{
The pipe will invalidate the instance if it is not able to find the
extension from the path of the instance.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to obtain the extension of instance.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preprocess.
}
}
}
}
}

\item{\bold{obtainExtension}}{
Getter of extension of the path.
\itemize{
\item{\emph{Usage}}{

\code{obtainExtension(path)}
}
\item{\emph{Value}}{

Extension of the path.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{path}}{
(character) Path of the file to get the extension.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
