% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StripHTMLPipe.R
\docType{class}
\name{StripHTMLPipe}
\alias{StripHTMLPipe}
\title{Class to remove html tags}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StripHTMLPipe$new(propertyName = "",
                  alwaysBeforeDeps = list(),
                  notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Class to remove html tags.
}
\details{
The pipe will invalidate the instance in the moment that the
resulting data is empty.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to remove html tags.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
}
}
}
}

\item{\bold{getDataWithOutHtml}}{
Function to get the data without html tags
\itemize{
\item{\emph{Usage}}{

\code{getDataWithOutHtml(data)}
}
\item{\emph{Value}}{

The data witouth html tags
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to remove html tags.
}
}
}
}
}

\item{\bold{cleanText}}{
Function to remove \\t,\\n and spaces from the text.
\itemize{
\item{\emph{Usage}}{

\code{cleanText(plainText)}
}
\item{\emph{Value}}{

The text without \\t,\\n and spaces.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{plainText}}{
(character) Text to remove \\t,\\n and spaces.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
