% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynsetFeatureVector.R
\docType{class}
\name{SynsetFeatureVector}
\alias{SynsetFeatureVector}
\title{Class to represent a vector of synset-based features}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SynsetFeatureVector$new(synsetFeature)
}
\arguments{
\item{synsetFeature}{(list) A map of synsets together with its values.}
}
\description{
Class to represent a vector of synset-based features.
}
\section{Methods}{

\itemize{
\item{\bold{getSynsetsFeature}}{
Getter of synsetFeature.
\itemize{
\item{\emph{Usage}}{

\code{getSynsetsFeature()}
}
\item{\emph{Value}}{

Value of synsetFeature.
}
}
}

\item{\bold{getSize}}{
Gets the size (number of properties) of the current SynsetFeatureVector.
\itemize{
\item{\emph{Usage}}{

\code{getSize()}
}
\item{\emph{Value}}{

The size of the current SynsetFeatureVector.
}
}
}

\item{\bold{getFrequencyValue}}{
Checks for the value stored for the synset synsetId.
\itemize{
\item{\emph{Usage}}{

\code{getFrequencyValue(synsetId)}
}
\item{\emph{Value}}{

The value asociated to synsetId, which represents the frequency of appearance
of the synsetId. If the synset is not found, -1 is returned.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsetId}}{
(character) The target synset.
}
}
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{synsetFeature}}{
 (list) A map of synsets together with its values
}
}
}

\keyword{datasets}
