% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynsetVector.R
\docType{class}
\name{SynsetVector}
\alias{SynsetVector}
\title{Class to represent a vector of synsets and the asociated information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SynsetVector$new(originalText)
}
\arguments{
\item{originalText}{(originalText) The start index of an entry.}
}
\description{
A class to represent a vector of synsets and the asociated information.
}
\section{Methods}{

\itemize{
\item{\bold{getOriginalText}}{
Getter of originalText.
\itemize{
\item{\emph{Usage}}{

\code{getOriginalText()}
}
\item{\emph{Value}}{

Value of originalText.
}
}
}

\item{\bold{setOriginalText}}{
Setter of originalText.
\itemize{
\item{\emph{Usage}}{

\code{setOriginalText(originalText)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{originalText}}{
(character) The new value of originalText.
}
}
}
}
}

\item{\bold{getFixedText}}{
Getter of fixedText.
\itemize{
\item{\emph{Usage}}{

\code{getFixedText()}
}
\item{\emph{Value}}{

Value of fixedText.
}
}
}

\item{\bold{setFixedText}}{
Setter of fixedText.
\itemize{
\item{\emph{Usage}}{

\code{setFixedText(fixedText)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{fixedText}}{
(character) The new value of fixedText.
}
}
}
}
}

\item{\bold{getUnmatchedTexts}}{
Getter of unmatchedTexts.
\itemize{
\item{\emph{Usage}}{

\code{getUnmatchedTexts()}
}
\item{\emph{Value}}{

Value of unmatchedTexts.
}
}
}

\item{\bold{setUnmatchedTexts}}{
Setter of unmatchedTexts.
\itemize{
\item{\emph{Usage}}{

\code{setUnmatchedTexts(unmatchedTexts)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{unmatchedTexts}}{
(list) The new value of unmatchedTexts.
}
}
}
}
}

\item{\bold{getSynsets}}{
Getter of synsets.
\itemize{
\item{\emph{Usage}}{

\code{getSynsets()}
}
\item{\emph{Value}}{

Value of synsets.
}
}
}

\item{\bold{setSynsets}}{
Setter of synsets.
\itemize{
\item{\emph{Usage}}{

\code{setSynsets(synsets)}
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsets}}{
(list) The new value of synsets.
}
}
}
}
}
}
}

\section{Private fields}{

\itemize{
\item{\bold{originalText}}{
 (character) The original text.
}
\item{\bold{fixedText}}{
 (character) The text after fixing unmatched text sections.
}
\item{\bold{unmatchedTexts}}{
 (list) The list of detected synsets represented as Pairs where:
}
\item{\bold{synsets}}{
 (list) The list of detected synsets represented as Pairs where:
}

}
}

\keyword{datasets}
