% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynsetVector2SynsetFeatureVectorPipe.R
\docType{class}
\name{SynsetVector2SynsetFeatureVectorPipe}
\alias{SynsetVector2SynsetFeatureVectorPipe}
\title{Class to transform a SynsetVector into a SynsetFeatureVector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SynsetVector2SynsetFeatureVectorPipe$new(propertyName = "",
                                         alwaysBeforeDeps = list(),
                                         notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
A pipe to transform a \code{SynsetVector} wich contains a list of
synsets included in a message into a \code{SynsetFeatureVector} wich compile togeher
duplicated features and assign a score for each feature according with a
groupingStrategy.
}
\details{
The groupStrategy is one of the following:
  - COUNT: indicates the number of times that a synset is observed in the
  content.
  - BOOLEAN: Indicates if the synset is observed in the content (1) or
  not (0).
  - FREQUENCY: Indicates the frequency of the synset in the text that is the
  count of times that the synset is observed divided by the whole amount of
  synsets.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to remove html tags.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, groupStrategy = "COUNT")}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{groupStrategy}}{
(character) Strategy to count the synstes. It can be COUNT, BOOLEAN or
FREQUENCY.
}
}
}
}
}

\item{\bold{countMatchers}}{
Converts a synsetVector in a synsetFeatureVector, with the synsetId and
the number of times that a synsetId appears in synsetVector.
\itemize{
\item{\emph{Usage}}{

\code{countMatchers(synsetVector)}
}
\item{\emph{Value}}{

A synsetFeatureVector with the synsetId and the number of times that a
synsetId appears in synsetVector.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsetVector}}{
(SynsetVector)
}
}
}
}
}

\item{\bold{booleanMatchers}}{
Converts a synsetVector in a synsetFeatureVector, with the synsetId and
1 or 0 depends if the synsetId appears in synsetVector.
\itemize{
\item{\emph{Usage}}{

\code{booleanMatchers(synsetVector)}
}
\item{\emph{Value}}{

A synsetFeatureVector with the synsetId and 1 or 0 depends if the synsetId
appears in synsetVector.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsetVector}}{
(SynsetVector)
}
}
}
}
}

\item{\bold{frequencyMatchers}}{
Converts a synsetVector in a synsetFeatureVector, with the synsetId and the
frequency that a synsetId appears in synsetVector.
\itemize{
\item{\emph{Usage}}{

\code{frequencyMatchers(synsetVector)}
}
\item{\emph{Value}}{

A synsetFeatureVector with the synsetId and the frequency that a synsetId
appears in synsetVector.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{synsetVector}}{
(SynsetVector)
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
