% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeeCSVPipe.R
\docType{class}
\name{TeeCSVPipe}
\alias{TeeCSVPipe}
\title{Class to complete the csv with the preprocessed instance}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TeeCSVPipe$new(propertyName = "",
               alwaysBeforeDeps = list(),
               notAfterDeps = list())
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Complete the csv with the preprocessed instance.
}
\details{
Building...
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that complete the csv with the preprocessed instance.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance, withData = TRUE, withSource = TRUE, outPutPath = "dataFrameAll.csv")}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
\item{\strong{withData}}{
(logical) Indicate if the data is added to csv.
}
\item{\strong{withSource}}{
(logical) Indicate if the source is added to csv.
}
\item{\strong{outPutPath}}{
(logical) Indicate if the source is added to csv
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
