% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToLowerCasePipe.R
\docType{class}
\name{ToLowerCasePipe}
\alias{ToLowerCasePipe}
\title{Class to convert the data of an instance to lower case}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ToLowerCasePipe$new(propertyName = "",
                    alwaysBeforeDeps = list(),
                    notAfterDeps = list("AbbreviationPipe", "SlangPipe"))
}
\arguments{
\item{propertyName}{(character) Name of the property associated with the pipe.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Class to convert the data to lowercase.
}
\section{Inherit}{

This class inherits from \code{\link{PipeGeneric}} and implements the
\code{pipe} abstract function.
}

\section{Methods}{

\itemize{
\item{\bold{pipe}}{
Function that preprocesses the instance to convert the data to lower case.
\itemize{
\item{\emph{Usage}}{

\code{pipe(instance)}
}
\item{\emph{Value}}{

The instance with the modifications that have occurred in the pipe.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) Instance to preproccess.
}
}
}
}
}

\item{\bold{toLowerCase}}{
Function that convert the data to lowercase.
\itemize{
\item{\emph{Usage}}{

\code{toLowerCase(data)}
}
\item{\emph{Value}}{

Data in lowercase.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{data}}{
(character) Text to preproccess.
}
}
}
}
}
}
}

\seealso{
\code{\link{PipeGeneric}}, \code{\link{Instance}}
}
\keyword{datasets}
