% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypePipe.R
\docType{class}
\name{TypePipe}
\alias{TypePipe}
\title{Class to establish the flow of pipes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TypePipe$new()
}
\description{
Class to establish the flow of pipes.
}
\details{
Building...
}
\section{Methods}{

\itemize{
\item{\bold{pipeAll}}{
Function where the flow of the pipes is created.
\itemize{
\item{\emph{Usage}}{

\code{pipeAll(instance)}
}
\item{\emph{Value}}{

The preprocessed instance.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{instance}}{
(Instance) The instance that is going to be processed.
}
}
}
}
}
}
}

\seealso{
\code{\link{Instance}}
}
\keyword{datasets}
