% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrbanDictionaryHandler.R
\docType{class}
\name{UrbanDictionaryHandler}
\alias{UrbanDictionaryHandler}
\title{Class to get the match with the specific word from an Urban dictionary}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UrbanDictionaryHandler$new(pathResourcesSlangs = "resources/slangs-json")
}
\arguments{
\item{propertyName}{(character) Name of the property.}

\item{alwaysBeforeDeps}{(list) The dependences alwaysBefore (pipes that must
be executed before this one).}

\item{notAfterDeps}{(list) The dependences notAfter (pipes that cannot be
executed after this one).}
}
\description{
Class to get the match with the specific word from an
Urban dictionary.
}
\section{Methods}{

\itemize{
\item{\bold{handle}}{
Get the matches with the originalText and set it to replacementText.
\itemize{
\item{\emph{Usage}}{

\code{handle(originalText, replacementText, lang)}
}
\item{\emph{Value}}{

The word that matches with the originalText.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{originalText}}{
(character)  The original text to replace.
}
\item{\strong{replacementText}}{
(character) The word that matches with the originalText.
}
\item{\strong{lang}}{
(character) The language of the original string.
}
}
}
}
}

\item{\bold{getReplacement4SlangTerm}}{
Get the matches with the originalText and set it to replacementText.
\itemize{
\item{\emph{Usage}}{

\code{getReplacement4SlangTerm(slangTerm, lang)}
}
\item{\emph{Value}}{

The word that matches with the originalText.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{slangTerm}}{
(character) The original text to replace.
}
\item{\strong{lang}}{
(character) The language of the original string.
}
}
}
}
}

\item{\bold{getPathResourcesSlangs}}{
Getter of path of slangs resources.
\itemize{
\item{\emph{Usage}}{

\code{getPathResourcesSlangs()}
}
\item{\emph{Value}}{

Value of path of slangs eesources.
}
}
}

}
}

\section{Private fields}{

\itemize{
\item{\bold{pathResourcesSlangs}}{
 (character) The path where are the resources.
}
}
}

\seealso{
\code{UnmatchedTextHandler}
}
\keyword{datasets}
