% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bdp4R.R
\docType{class}
\name{Bdp4R}
\alias{Bdp4R}
\title{Class to manage the preprocess of the files through the pipes' flow}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Bdp4R$new(pathKeys = "config/configurations.ini")
}
\arguments{
\item{connections}{(Connections) Initialize the object that handles the
different types of connections with youtube and twitter.}

\item{babelUtils}{(BabelUtils) Initialize the object that handles the different
types of connections with babelfy and babelnet.}

\item{resourceHandle}{(ResourceHandler) Initialize the object that manages the
loading of the resource files, such as abbreviation, slang, stopword, etc.}

\item{dataFrameAll}{(data.frame) Variable in which the data.frame generated
by the pipes will be saved.}

\item{dataFrameAllSynsets}{(data.frame) Variable in which a data.frame is
built with the instance synsets.}
}
\description{
It contains the "static" variables that will be used throughout
the classes and the function that prepares and launches the execution of the
pipes from the TypePipe object that is passed to it as an argument. Also,
build the csvs / RDatas from the static variables of type data.frame,
in this case, dataFrameAll and dataFrameAllSynsets.
}
\section{Methods}{

\itemize{
\item{\bold{proccess_files}}{
Preprocess files through a pipes' flow
\itemize{
\item{\emph{Usage}}{

\code{proccess_files(pathFiles,
                     pipe,
                     pathOutPut = "output.RData",
                     pathOutPutSynsets = "outputSynsets.RData")}
}
\item{\emph{Value}}{

List of instances that have been preprocessed.
}
\item{\emph{Arguments}}{
\itemize{
\item{\strong{pathFiles}}{
(character) Path where the files to be processed are located.
}
\item{\strong{pipe}}{
(TypePipe) Indicate if the abbreviations are replaced.
}
\item{\strong{pathOutPut}}{
(character) Path of the file where the csv or dataFrame of the processed
instances will be saved without the synsets.
}
\item{\strong{pathOutPutSynsets}}{
(character) Path of the file where the csv or dataFrame of the processed
instances will be saved with the synsets.
}
}
}
}
}
}
}

\keyword{datasets}
