% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdp4R_execute.R
\docType{methods}
\name{bdp4R_execute}
\alias{bdp4R_execute}
\title{Preproccess files through pipes}
\usage{
bdp4R_execute(configurationFilePath, filesPath, pipe = SerialPipes$new())
}
\arguments{
\item{configurationFilePath}{(character) Path where the file with keys are located.}

\item{filesPath}{(character) Path where the files to be preprocessed are located.}

\item{pipe}{(TypePipe) Subclass of TypePipe, which implements the pipe method.}
}
\value{
List of instances that have been preprocessed.
}
\description{
Method that allows to preprocess files in a comfortable way.
}
